fix(number-input): fix TS error
Makes `.value` property of `<bx-input>` an accessor so
`<bx-number-input>`, that inherits `<bx-input>`, doesn't have to
redefine the type.

Refs #576.
