+++ src/components/number-input/number-input.ts
@@ -1,13 +1,13 @@
 /**
  * @license
  *
- * Copyright IBM Corp. 2019, 2020
+ * Copyright IBM Corp. 2019, 2021
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
 
-import { customElement, html, property, query } from 'lit-element';
+import { customElement, html, property } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map';
 import settings from 'carbon-components/es/globals/js/settings';
 import WarningFilled16 from '@carbon/icons/lib/warning--filled/16';
@@ -31,12 +31,6 @@ const { prefix } = settings;
  */
 @customElement(`${prefix}-number-input`)
 export default class BXNumberInput extends BXInput {
-  /**
-   * The underlying input element
-   */
-  @query('input')
-  protected _input!: HTMLInputElement;
-
   _testValidity() {
     if (this._input?.valueAsNumber > Number(this.max)) {
       return NUMBER_INPUT_VALIDATION_STATUS.EXCEEDED_MAXIMUM;
@@ -63,8 +57,6 @@ export default class BXNumberInput extends BXInput {
   protected _max = '';
 
   protected _step = '1';
-
-  protected _value = '';
 
   /**
    * The color scheme.
@@ -112,33 +104,6 @@ export default class BXNumberInput extends BXInput {
     const oldValue = this.step;
     this._step = value;
     this.requestUpdate('step', oldValue);
-  }
-
-  /**
-   * The value of the input.
-   */
-  @property({ reflect: true })
-  // @ts-ignore
-  get value() {
-    // FIXME: Figure out how to deal with TS2611
-    // once we have the input we can directly query for the value
-    if (this._input) {
-      return this._input.value;
-    }
-    // but before then _value will work fine
-    return this._value;
-  }
-
-  set value(value) {
-    const oldValue = this._value;
-    this._value = value;
-    // make sure that lit-element updates the right properties
-    this.requestUpdate('value', oldValue);
-    // we set the value directly on the input (when available)
-    // so that programatic manipulation updates the UI correctly
-    if (this._input) {
-      this._input.value = value;
-    }
   }
 
   /**
